+++ src/plonemeeting/restapi/serializer/summary.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
-from plone.restapi.interfaces import ISerializeToJsonSummary
 from imio.restapi.serializer.base import DefaultJSONSummarySerializer
+from plone.restapi.interfaces import ISerializeToJsonSummary
+from plonemeeting.restapi.serializer.base import BaseSerializeToJson
 from Products.ZCatalog.interfaces import ICatalogBrain
 from OFS.interfaces import IItem
 from zope.component import adapter
@@ -11,16 +12,8 @@ from zope.interface import Interface
 
 @implementer(ISerializeToJsonSummary)
 @adapter(ICatalogBrain, Interface)
-class PMBrainJSONSummarySerializer(DefaultJSONSummarySerializer):
+class PMBrainJSONSummarySerializer(DefaultJSONSummarySerializer, BaseSerializeToJson):
     """ISerializeToJsonSummary adapter for brain."""
-
-    def _extra_include(self, result):
-        """ """
-        return result
-
-    def _additional_values(self, result):
-        """ """
-        return result
 
     @property
     def _additional_fields(self):
@@ -30,15 +23,7 @@ class PMBrainJSONSummarySerializer(DefaultJSONSummarySerializer):
     def __call__(self):
         """ """
         result = super(PMBrainJSONSummarySerializer, self).__call__()
-
-        # fullobjects for extra_includes?
-        self.extra_include_fullobjects = False
-        if "extra_include_fullobjects" in self.request.form:
-            self.extra_include_fullobjects = True
-
-        result = self._extra_include(result)
-        result = self._additional_values(result)
-
+        result = self._after__call__(result)
         return result